requirements.txt
binance-futures-connector
pandas-ta
discord-webhook
mplfinance